+++ predictive_punter/__init__.py
@@ -1,14 +1,16 @@
 __version__ = '1.0.0a5'
 
 
-from . import race
-from . import runner
 from .sample import Sample
 from .predictor import Predictor
 from .prediction import Prediction
 from .provider import Provider
 
+from . import race
+from . import runner
+
 from .command import Command
 from .scrape import ScrapeCommand
 from .seed import SeedCommand
 from .simulate import SimulateCommand
+from .predict import PredictCommand